move saml_helpers.py into the routes folder
